requirements.txt
asgiref==3.10.0
Django==5.2.7
django-mssql-backend==2.8.1
djangorestframework==3.16.1
mssql-django==1.6
pyodbc==5.3.0
python-dotenv==1.1.1
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
